+++ apps/web/src/views/Swap/MMLinkPools/components/MMCommitButton.tsx
@@ -13,7 +13,7 @@ import { Field } from 'state/swap/actions'
 import { logGTMClickSwapEvent } from 'utils/customGTMEventTracking'
 import { Address } from 'viem'
 import { parseMMError } from 'views/Swap/MMLinkPools/utils/exchange'
-import { ConfirmSwapModal } from 'views/Swap/V3Swap/containers/ConfirmSwapModalV2'
+import { ConfirmSwapModal } from 'views/Swap/V3Swap/containers/ConfirmSwapModal'
 import { CommitButtonProps } from 'views/Swap/V3Swap/types'
 import { MMTradeInfo } from '../hooks'
 import { useMMConfirmModalState } from '../hooks/useMMConfirmModalState'
