+++ apps/web/src/views/Swap/MMLinkPools/hooks/useMMConfirmModalState.ts
@@ -9,7 +9,7 @@ import { RetryableError, retry } from 'state/multicall/retry'
 import { UserUnexpectedTxError } from 'utils/errors'
 import { publicClient } from 'utils/wagmi'
 import { Address, Hex, TransactionNotFoundError, TransactionReceipt, TransactionReceiptNotFoundError } from 'viem'
-import { ConfirmAction } from 'views/Swap/V3Swap/hooks/useConfirmModalStateV2'
+import { ConfirmAction } from 'views/Swap/V3Swap/hooks/useConfirmModalState'
 import { userRejectedError } from 'views/Swap/V3Swap/hooks/useSendSwapTransaction'
 import { useApprove, useApproveRequires } from './useApprove'
 import { MMSwapCall } from './useSwapCallArguments'
