+++ apps/web/src/views/Swap/V3Swap/hooks/index.ts
@@ -4,5 +4,5 @@ export * from './useParsedAmounts'
 export * from './useRouterAddress'
 export * from './useSlippageAdjustedAmounts'
 export * from './useSwapBestTrade'
-export * from './useSwapCallbackV2'
+export * from './useSwapCallback'
 export * from './useSwapInputError'
